Update what-is-iam.md (#267)
In general, I like the declarative style of the guide, however in this case, the sentence in isolation suggests a ill advised solution without preparing the reader to realize that it is a negative example. Changing this from "This can be done" to "While this could be done" alerts the reader up front to be wary of the statement, and the conjunction to attach it to the next sentence explains why.